agents/strands_agent/requirements-agentcore.txt
bedrock-agentcore>=1.3.1
strands-agents>=1.27.0
boto3>=1.40.0
python-dotenv>=1.0.0